Refactors collection creation into reusable component
